pyproject.toml
[project]
name = "etf-quant-assistant"
version = "0.1.0"
description = "ETF量化助手桌面应用"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "akshare>=1.11.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
